+++ config/initializers/elasticsearch.rb
@@ -3,6 +3,7 @@
 require 'elasticsearch/model'
 
 Elasticsearch::Model.client = Elasticsearch::Client.new(
-  host: ENV.fetch('ELASTICSEARCH_ADDRESS', nil)
+  host: ENV.fetch('ELASTICSEARCH_ADDRESS', nil),
+  adapter: :net_http
 )
 Hashie.logger = Logger.new('/dev/null')