+++ CHANGELOG.md
@@ -4,12 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/learningtapestry/lcms-engine/compare/v0.5.4...HEAD)
+## [Unreleased](https://github.com/learningtapestry/lcms-engine/compare/v0.5.5...HEAD)
+
+
+## [0.5.5](https://github.com/learningtapestry/lcms-engine/compare/v0.5.4...v0.5.5) - 2023.12.24
 
 ### Changed
 
 - Enhance error handling for job result fetching - [@paranoicsan](https://github.com/paranoicsan)
 - Normalize metadata search in Resource model - [@paranoicsan](https://github.com/paranoicsan)
+- Bumped lt-lcms/lt-google-api to use upgraded google api (0.11) - [@shlag3n](https://github.com/shlag3n)
 
 ## [0.5.4](https://github.com/learningtapestry/lcms-engine/compare/v0.5.3...v0.5.4) - 2022.12.30